Pull in PouchDB as demo vehicle
PouchDB should be simpler to work with as it's all local, to work out
simple things like how to AND efficiently, and OR, and work out the
"basic operations" on the index, to encode into methods. But for now
we just have a couple of trivial queries to start us on our way.