Emit fields in index as objects
Some document schema create issues with ranges. Take the following contrived schema:

```
{"id": "d", "username": "mike", "role": "editor"}
{"id": "e", "username": "rob", "role": "editor"}
{"id": "f", "username": {"mike": "foo" } }
```

This will create the index keys:

```
["username", "mike"] -> d
["username", "mike", "foo"] -> f
["username", "rob"] -> e
```

Using the above, the query `username > mike` "should" return just `e`.
But if we work it through we can see that it would not. Instead it
returns both f and e.

To alleviate this we need to differentiate keys from values in our
index. We need something that "sorts after" all possible values for
a field. For CouchDB collation, an object satisfies that condition.

So we index as:

```
[`{"username": null}`, "mike"] -> d
[`{"username": null}`, "rob"] -> e
[`{"username": null}`, `{"mike": null}`, "foo"] -> e
```

Our range query `startkey = ["username", "mike"],
endkey = ["username", {}]` now stops at the field boundary
and returns only `e`, as expected.